// src/types/travel.ts
export type TravelLocation = {
  latitude: string;
  longitude: string;
  locationName: string;
  subdistrict: string;
  district: string;
  province: string;
};

export type UploadedPhoto = {
  driveFileId: string;
  driveUrl: string;
  thumbnailUrl: string;
  fileName: string;
  mimeType: string;
  previewDriveFileId?: string;
};